refactor + update code text color
