+++ chat.css
@@ -23,7 +23,7 @@ InputText {
 ResponseCode {
     width: 100%;
     margin: 0 0 1 25;
-    color: #b85616;
+    color: #ee7e28;
 }
 
 ResponseText {
